+++ src/hsp3/emscripten/supio_emscripten.h
@@ -8,10 +8,10 @@
 
 size_t utf8strlen(const char *target);
 
-char *mem_ini( int size );
+char *mem_ini( size_t size );
 void mem_bye( void *ptr );
 int mem_save( char *fname, void *mem, int msize, int seekofs );
-char *mem_alloc( void *base, int newsize, int oldsize );
+char *mem_alloc( void *base, size_t newsize, size_t oldsize );
 
 void strcase( char *str );
 int strcpy2( char *str1, char *str2 );
