+++ src/hsp3/linux/supio_linux.cpp
@@ -151,7 +151,7 @@ static int wildcard( char *text, char *wc )
 //
 static FILE *fp;
 
-char *mem_ini( int size ) {
+char *mem_ini( size_t size ) {
 	return (char *)calloc(size,1);
 }
 
@@ -885,7 +885,7 @@ void Alertf( char *format, ... )
 //
 //		Memory Manager
 //
-char *mem_alloc( void *base, int newsize, int oldsize )
+char *mem_alloc( void *base, size_t newsize, size_t oldsize )
 {
 	char *p;
 	if ( base == NULL ) {
