+++ src/hsp3/win32gui/supio_win_unicode.cpp
@@ -28,7 +28,7 @@
 //
 static FILE *fp;
 
-char *mem_ini( int size ) {
+char *mem_ini(size_t size ) {
 	return (char *)calloc(size,1);
 }
 
@@ -36,7 +36,7 @@ void mem_bye( void *ptr ) {
 	free(ptr);
 }
 
-char *mem_alloc( void *base, int newsize, int oldsize )
+char *mem_alloc( void *base, size_t newsize, size_t oldsize )
 {
 	char *p;
 	if ( base == NULL ) {
